requirements.txt
Flask
requests
Python-dotenv